requirements.txt
streamlit
Pillow
ultralytics
pandas
